+++ R/opts.R
@@ -220,6 +220,7 @@ get_colorized_params <- function(
 #' - `ihtml.page_size_values`
 #' - `ihtml.pagination_type`
 #' - `ihtml.height`
+#' - `ihtml.selection_mode`
 #'
 #' @inheritParams fmt_number
 #'
@@ -337,6 +338,17 @@ get_colorized_params <- function(
 #'
 #'   Height of the table in pixels. Defaults to `"auto"` for automatic sizing.
 #'
+#' @param selection_mode *Allow row selection*
+#'
+#'    `scalar<character>` // *default:* `NULL`
+#'
+#'    The `selection_mode` options allows users to select rows by clicking them.
+#'    When this option is `"single"`, clicking another value toggles selection
+#'    of the previously selected row off. When this option is `"multiple"`,
+#'    multiple rows can be selected at once. Selected values are available in
+#'    Shiny apps when `selection_mode` is not `NULL` and the table is used in
+#'    [render_gt()].
+#'
 #' @return An object of class `gt_tbl`.
 #'
 #' @section Examples:
@@ -418,13 +430,14 @@ opt_interactive <- function(
     page_size_default = 10,
     page_size_values = c(10, 25, 50, 100),
     pagination_type = c("numbers", "jump", "simple"),
-    height = "auto"
+    height = "auto",
+    selection_mode = NULL
 ) {
 
   # Perform input object validation
   stop_if_not_gt_tbl(data = data)
 
-  pagination_type <- 
+  pagination_type <-
     rlang::arg_match0(pagination_type, values = c("numbers", "jump", "simple"))
 
   tab_options(
@@ -443,7 +456,8 @@ opt_interactive <- function(
     ihtml.page_size_default = page_size_default,
     ihtml.page_size_values = page_size_values,
     ihtml.pagination_type = pagination_type,
-    ihtml.height = height
+    ihtml.height = height,
+    ihtml.selection_mode = selection_mode
   )
 }
 
@@ -1457,11 +1471,11 @@ opt_table_outline <- function(
 #'   A name that is representative of a font stack (obtained via internally via
 #'   the [system_fonts()] helper function). If provided, this new stack will
 #'   replace any defined fonts and any `font` values will be prepended.
-#' 
+#'
 #' @param size *Text size*
-#' 
+#'
 #'   `scalar<character|numeric|integer>` // *default:* `NULL` (`optional`)
-#' 
+#'
 #'   The text size for the entire table can be set by providing a `size` value.
 #'   Can be specified as a single-length character vector with units of pixels
 #'   (e.g., `12px`) or as a percentage (e.g., `80%`). If provided as a
@@ -1484,11 +1498,11 @@ opt_table_outline <- function(
 #'   `"normal"`, `"bold"`, `"lighter"`, `"bolder"`, or, a numeric value between
 #'   `1` and `1000`, inclusive. Please note that typefaces have varying support
 #'   for the numeric mapping of weight.
-#' 
+#'
 #' @param color *Text color*
-#' 
+#'
 #'   `scalar<character>` // *default:* `NULL` (`optional`)
-#' 
+#'
 #'   The `color` option defines the text color used throughout the table. A
 #'   color name or a hexadecimal color code should be provided.
 #'